fix invitation process
